requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pymysql
python-dotenv
pydantic
pydantic-settings
passlib[bcrypt]
pydantic[email]
python-jose[cryptography]
bcrypt==4.0.1